pyproject.toml
[tool.poetry]
name = "tech-code-rv"
version = "0.1.0"
description = ""
authors = ["Lucas Vizentin <Lucas.Vizentin@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
factory-boy = "^3.3.0"
django = "^5.0.6"
djangorestframework = "^3.15.1"
requests = "^2.32.3"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
